requirement.txt
numpy
matplotlib
pytorch
scikit-learn